showtext_end {showtext}	R Documentation

Turning Off 'showtext' Text Rendering

Description

The two versions of this function are equivalent, but the "underscore" naming is preferred.

This function will turn off the showtext functionality of rendering text. When you call this function, the current active device should be the same as the one when you call showtext_begin(), or an error will be issued. See the example in showtext_begin().

Usage

showtext_end()

showtext.end()

Author(s)

Yixuan Qiu <https://statr.me/>

See Also

showtext_begin()

[Package showtext version 0.9-2 Index]